Added new Encoder class
This allows us to convert messages to binary and force them to
always result in printable characters when passed through the Encryption
function.

Also, adjusted the nets encryptPlaintext and decryptBob/Alice functions
to accept string inputs like 'testmessage'

If the message is a multiple of the network word size, then it encrypts
or decrypts block-by-block

And finally, set a seed for the weight initializer that seems to usually
produce good results as long as the learning rate isn't too small.